+++ src/streams.rs
@@ -1,4 +1,5 @@
 //  Library
+use crate::ansi;
 use crate::config;
 use crate::stream::Stream;
 use crate::utils;
@@ -17,14 +18,16 @@ impl Streams {
         min_count: u16,
         max_count: u16,
         mode: utils::Mode,
-        color: utils::RGBColor,
+        color: ansi::RGBColor,
     ) -> Self {
         //  Instantiate streams vector
         let mut streams: Vec<Stream> = Vec::new();
 
         //  Generate stream entities
         for c in 0..count {
-            if c % config::STREAM_SPACING != 0 { continue }
+            if c % config::STREAM_SPACING != 0 {
+                continue;
+            }
             //  Generate a stream
             let height_offset = utils::random_between(-50, 0);
             let stream = Stream::new(c as i32, height_offset, min_count, max_count, color, mode);
